#3 Refactor: refactor dto directory
