+++ src/main/java/kahlua/KahluaProject/config/CorsMvcConfig.java
@@ -1,4 +1,4 @@
-package kahlua.KahluaProject.confg;
+package kahlua.KahluaProject.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
