+++ src/main/java/kahlua/KahluaProject/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package kahlua.KahluaProject.confg;
+package kahlua.KahluaProject.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
