+++ src/main/java/kahlua/KahluaProject/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package kahlua.KahluaProject.confg;
+package kahlua.KahluaProject.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Info;
