+++ src/main/java/kahlua/KahluaProject/controller/AdminController.java
@@ -5,11 +5,11 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import kahlua.KahluaProject.apipayload.ApiResponse;
 import kahlua.KahluaProject.apipayload.code.status.ErrorStatus;
 import kahlua.KahluaProject.domain.user.UserType;
-import kahlua.KahluaProject.dto.response.ApplyGetResponse;
-import kahlua.KahluaProject.dto.response.ApplyListResponse;
+import kahlua.KahluaProject.dto.apply.response.ApplyGetResponse;
+import kahlua.KahluaProject.dto.apply.response.ApplyListResponse;
 import kahlua.KahluaProject.exception.GeneralException;
 import kahlua.KahluaProject.security.AuthDetails;
-import kahlua.KahluaProject.dto.response.TicketListResponse;
+import kahlua.KahluaProject.dto.ticket.response.TicketListResponse;
 import kahlua.KahluaProject.service.ApplyService;
 import kahlua.KahluaProject.service.TicketService;
 import lombok.RequiredArgsConstructor;
