+++ src/main/java/kahlua/KahluaProject/controller/AuthController.java
@@ -7,9 +7,9 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import kahlua.KahluaProject.apipayload.ApiResponse;
 import kahlua.KahluaProject.apipayload.code.status.ErrorStatus;
-import kahlua.KahluaProject.dto.request.SignInRequest;
-import kahlua.KahluaProject.dto.request.SignUpRequest;
-import kahlua.KahluaProject.dto.response.TokenResponse;
+import kahlua.KahluaProject.dto.user.request.SignInRequest;
+import kahlua.KahluaProject.dto.user.request.SignUpRequest;
+import kahlua.KahluaProject.dto.user.response.TokenResponse;
 import kahlua.KahluaProject.exception.GeneralException;
 import kahlua.KahluaProject.security.AuthDetails;
 import kahlua.KahluaProject.security.jwt.JwtProvider;
