+++ src/main/java/kahlua/KahluaProject/converter/AuthConverter.java
@@ -1,8 +1,8 @@
 package kahlua.KahluaProject.converter;
 
 import kahlua.KahluaProject.domain.user.User;
-import kahlua.KahluaProject.dto.response.SignInResponse;
-import kahlua.KahluaProject.dto.response.TokenResponse;
+import kahlua.KahluaProject.dto.user.response.SignInResponse;
+import kahlua.KahluaProject.dto.user.response.TokenResponse;
 
 public class AuthConverter {
     public static SignInResponse toSignInResDto(User user, TokenResponse tokenResponse) {
