+++ src/main/java/kahlua/KahluaProject/dto/apply/response/ApplyItemResponse.java
@@ -1,4 +1,4 @@
-package kahlua.KahluaProject.dto.response;
+package kahlua.KahluaProject.dto.apply.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import kahlua.KahluaProject.domain.apply.Gender;
