+++ src/main/java/kahlua/KahluaProject/dto/ticket/request/TicketCreateRequest.java
@@ -1,28 +1,23 @@
-package kahlua.KahluaProject.dto.response;
+package kahlua.KahluaProject.dto.ticket.request;
 
 import kahlua.KahluaProject.domain.ticket.Meeting;
-import kahlua.KahluaProject.domain.ticket.Participants;
-import kahlua.KahluaProject.domain.ticket.Status;
 import kahlua.KahluaProject.domain.ticket.Type;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.util.List;
 
 @Getter
 @AllArgsConstructor
-@Builder
-public class TicketCreateResponse {
+@NoArgsConstructor
+public class TicketCreateRequest {
 
-    private Long id;
     private String buyer;
     private String phone_num;
-    private String reservationId;
     private Type type;
     private String major;
     private String studentId;
     private Meeting meeting;
-    private List<Participants> members;
-    private Status status;
+    private List<ParticipantsCreateRequest> members;
 }
