+++ src/main/java/kahlua/KahluaProject/dto/ticket/response/TicketCreateResponse.java
@@ -1,24 +1,28 @@
-package kahlua.KahluaProject.dto.request;
+package kahlua.KahluaProject.dto.ticket.response;
 
 import kahlua.KahluaProject.domain.ticket.Meeting;
 import kahlua.KahluaProject.domain.ticket.Participants;
+import kahlua.KahluaProject.domain.ticket.Status;
 import kahlua.KahluaProject.domain.ticket.Type;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 import java.util.List;
 
 @Getter
 @AllArgsConstructor
-@NoArgsConstructor
-public class TicketCreateRequest {
+@Builder
+public class TicketCreateResponse {
 
+    private Long id;
     private String buyer;
     private String phone_num;
+    private String reservationId;
     private Type type;
     private String major;
     private String studentId;
     private Meeting meeting;
-    private List<ParticipantsCreateRequest> members;
+    private List<Participants> members;
+    private Status status;
 }
