+++ src/main/java/kahlua/KahluaProject/dto/user/request/SignInRequest.java
@@ -1,4 +1,4 @@
-package kahlua.KahluaProject.dto.request;
+package kahlua.KahluaProject.dto.user.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
