+++ src/main/java/kahlua/KahluaProject/service/ApplyService.java
@@ -6,11 +6,11 @@ import kahlua.KahluaProject.converter.ApplyConverter;
 import kahlua.KahluaProject.domain.apply.Apply;
 import kahlua.KahluaProject.domain.user.User;
 import kahlua.KahluaProject.domain.user.UserType;
-import kahlua.KahluaProject.dto.request.ApplyCreateRequest;
-import kahlua.KahluaProject.dto.response.ApplyCreateResponse;
-import kahlua.KahluaProject.dto.response.ApplyGetResponse;
-import kahlua.KahluaProject.dto.response.ApplyItemResponse;
-import kahlua.KahluaProject.dto.response.ApplyListResponse;
+import kahlua.KahluaProject.dto.apply.request.ApplyCreateRequest;
+import kahlua.KahluaProject.dto.apply.response.ApplyCreateResponse;
+import kahlua.KahluaProject.dto.apply.response.ApplyGetResponse;
+import kahlua.KahluaProject.dto.apply.response.ApplyItemResponse;
+import kahlua.KahluaProject.dto.apply.response.ApplyListResponse;
 import kahlua.KahluaProject.exception.GeneralException;
 import kahlua.KahluaProject.repository.ApplyRepository;
 import lombok.RequiredArgsConstructor;
