+++ src/main/java/kahlua/KahluaProject/service/AuthService.java
@@ -5,11 +5,11 @@ import kahlua.KahluaProject.converter.AuthConverter;
 import kahlua.KahluaProject.converter.UserConverter;
 import kahlua.KahluaProject.domain.user.Credential;
 import kahlua.KahluaProject.domain.user.User;
-import kahlua.KahluaProject.dto.request.SignInRequest;
-import kahlua.KahluaProject.dto.request.SignUpRequest;
-import kahlua.KahluaProject.dto.response.SignInResponse;
-import kahlua.KahluaProject.dto.response.TokenResponse;
-import kahlua.KahluaProject.dto.response.UserResponse;
+import kahlua.KahluaProject.dto.user.request.SignInRequest;
+import kahlua.KahluaProject.dto.user.request.SignUpRequest;
+import kahlua.KahluaProject.dto.user.response.SignInResponse;
+import kahlua.KahluaProject.dto.user.response.TokenResponse;
+import kahlua.KahluaProject.dto.user.response.UserResponse;
 import kahlua.KahluaProject.exception.GeneralException;
 import kahlua.KahluaProject.redis.RedisClient;
 import kahlua.KahluaProject.security.jwt.JwtProvider;
