+++ src/main/java/kahlua/KahluaProject/service/TicketService.java
@@ -8,8 +8,11 @@ import kahlua.KahluaProject.domain.ticket.Ticket;
 import kahlua.KahluaProject.domain.ticket.Type;
 import kahlua.KahluaProject.domain.user.User;
 import kahlua.KahluaProject.domain.user.UserType;
-import kahlua.KahluaProject.dto.request.TicketCreateRequest;
-import kahlua.KahluaProject.dto.response.*;
+import kahlua.KahluaProject.dto.ticket.request.TicketCreateRequest;
+import kahlua.KahluaProject.dto.ticket.response.TicketCreateResponse;
+import kahlua.KahluaProject.dto.ticket.response.TicketGetResponse;
+import kahlua.KahluaProject.dto.ticket.response.TicketItemResponse;
+import kahlua.KahluaProject.dto.ticket.response.TicketListResponse;
 import kahlua.KahluaProject.exception.GeneralException;
 import kahlua.KahluaProject.repository.ParticipantsRepository;
 import kahlua.KahluaProject.repository.TicketRepository;
